feat(auth): Use better-auth API for admin actions and loaders
