+++ app/server.ts
@@ -1,15 +1,23 @@
 import {createHonoServer} from "react-router-hono-server/node"
 
+import type {Context} from "hono"
 import {csrf} from "hono/csrf"
 
-import {Auth} from "./server/lib/auth/Auth.js"
 import {orm} from "./server/lib/db/orm.js"
 import {withAuth} from "./server/middlewares/withAuth.js"
 import {withOrm} from "./server/middlewares/withOrm.js"
+import type {auth} from "./server/lib/auth/auth.js"
 
 import config from "./server/lib/config.js"
 
-export default await createHonoServer({
+export interface Variables {
+  orm: typeof orm
+  auth: typeof auth
+}
+
+export default await createHonoServer<{
+  Variables: Variables
+}>({
   port: config.server.port,
   async configure(hono) {
     await orm.connect()
@@ -20,10 +28,10 @@ export default await createHonoServer({
       .use(withAuth())
   },
 
-  async getLoadContext(ctx) {
+  async getLoadContext(ctx: Context<{Variables: Variables}>) {
     return {
-      auth: new Auth(ctx),
-      orm
+      auth: ctx.var.auth,
+      orm: ctx.var.orm
     }
   },
