+++ app/server/lib/admin/SharedAdminContext.ts
@@ -1,9 +1,3 @@
-import type {MikroORM} from "@mikro-orm/mariadb"
-import type {Context} from "hono"
+import type {Variables} from "../../../server.js"
 
-import type {Auth} from "../auth/Auth.js"
-
-export interface SharedAdminContext {
-  orm: MikroORM
-  auth: Auth<Context>
-}
+export type SharedAdminContext = Variables
