--- app/server/lib/auth/ttl.ts
@@ -1,3 +1,0 @@
-import {TimeSpan} from "lucia"
-
-export const ttl = new TimeSpan(30, "d")
